+++ CHANGELOG.md
@@ -5,6 +5,7 @@ To install or update LODA, please follow the [installation instructions](https:/
 ### Bugfixes
 
 * Optimizer does not consider `$1` as reserved cell anymore (formerly output cell)
+* Validator does not remove operations anymore that are needed for computing more terms
 
 ### Features